+++ src/atoms/FlatButton.jsx
@@ -23,23 +23,22 @@ const colorTypes = {
   },
 };
 
-const sizes = {
-  XS: typography.sizeXS,
-  S: typography.sizeS,
-  M: typography.sizeM,
-  L: typography.sizeL,
-  XL: typography.sizeXL,
-  XXL: typography.sizeXXL,
-};
-
-const FlatButton = ({ text, onClick, isDisabled, type, size }) => {
+const FlatButton = ({
+  buttonType,
+  text,
+  onClick,
+  isDisabled,
+  type,
+  size,
+  weight,
+}) => {
   const colorType = isDisabled ? colorTypes.disabled : colorTypes[type];
 
   const container = css`
     color: ${colorType.color};
     font-family: 'Open Sans', sans-serif;
-    font-weight: ${typography.semibold};
-    font-size: ${sizes[size]};
+    font-weight: ${typography[weight]};
+    font-size: ${typography[size]};
     text-decoration: underline;
     cursor: ${isDisabled ? 'default' : 'pointer'};
 
@@ -49,8 +48,10 @@ const FlatButton = ({ text, onClick, isDisabled, type, size }) => {
   `;
 
   return (
+    // eslint doesn't allow dynamic button type
+    // eslint-disable-next-line react/button-has-type
     <button
-      type="button"
+      type={buttonType}
       css={container}
       onClick={onClick}
       disabled={isDisabled}
@@ -64,19 +65,26 @@ FlatButton.propTypes = {
   /** Indicates the `<button>` is disabled */
   isDisabled: PropTypes.bool,
   /** Callback onClick */
-  onClick: PropTypes.func.isRequired,
+  onClick: PropTypes.func,
   /** Texto del botón */
   text: PropTypes.string.isRequired,
   /** Tipo de color del `<button>` */
   type: PropTypes.oneOf(['accent', 'disabled', 'info', 'error']),
   /** Font size */
   size: PropTypes.oneOf(['XS', 'S', 'M', 'L', 'XL', 'XXL']),
+  /** Font weight */
+  weight: PropTypes.oneOf(['bold', 'semibold', 'regular']),
+  /** Seelects button type */
+  buttonType: PropTypes.oneOf(['button', 'submit']),
 };
 
 FlatButton.defaultProps = {
+  onClick: () => {},
   type: 'accent',
   size: 'M',
+  weight: 'semibold',
   isDisabled: false,
+  buttonType: 'button',
 };
 
 export default FlatButton;