+++ src/tokens/typography.js
@@ -1,9 +1,9 @@
-export const sizeXXL = '20px';
-export const sizeXL = '18px';
-export const sizeL = '16px';
-export const sizeM = '15px';
-export const sizeS = '13px';
-export const sizeXS = '11px';
+export const XXL = '20px';
+export const XL = '18px';
+export const L = '16px';
+export const M = '15px';
+export const S = '13px';
+export const XS = '11px';
 
 export const bold = 700;
 export const semibold = 600;