+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
-## v0.3.0
+## v0.3.1
 
-- Adds variant that allows choose price or text on right side for ListItem
+##### Spacing component
+
+- Updates stories with use cases
+- Adds `baseline` on `align-items` set values
+
+##### Visibility
+
+- Adds visibility table to story docs
+
+##### General
+
+- Refactors `size` and `weight` constants using existing `typography` constants
+
+##### Button & FlatButton
+
+- Adds default value for onClick prop
+- Adds buttonType prop (warning alert)
+
+##### Link & FlatButton
+
+- Adds weight prop
+
+##### Text
+
+- Adds `white` on `color` set values